[cocas] Move source files to modules inside cocas
