+++ cocas/assembler/__init__.py
@@ -1,0 +1,5 @@
+from . import ast_nodes
+from .assembler import assemble
+from .ast_builder import build_ast
+from .code_block import CodeBlock, Section
+from .macro_processor import process_macros
