+++ cocas/assembler/assembler.py
@@ -2,13 +2,12 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Type
 
-from cocas.abstract_code_segments import CodeSegmentsInterface
-from cocas.abstract_instructions import TargetInstructionsInterface
-from cocas.ast_nodes import InstructionNode, LabelDeclarationNode, ProgramNode, TemplateSectionNode
-from cocas.code_block import Section
 from cocas.error import CdmExceptionTag
-from cocas.location import CodeLocation
-from cocas.object_module import ObjectModule, ObjectSectionRecord
+from cocas.object_module import CodeLocation, ObjectModule, ObjectSectionRecord
+from cocas.targets import CodeSegmentsInterface, TargetInstructionsInterface
+
+from .ast_nodes import InstructionNode, LabelDeclarationNode, ProgramNode, TemplateSectionNode
+from .code_block import Section
 
 TAG = CdmExceptionTag.ASM
