+++ cocas/object_file/__init__.py
@@ -1,0 +1,1 @@
+from .object_file import import_object
