+++ cocas/targets/__init__.py
@@ -1,0 +1,3 @@
+from .abstract_code_segments import CodeSegmentsInterface
+from .abstract_instructions import TargetInstructionsInterface
+from .abstract_params import TargetParamsInterface
