+++ cocas/targets/cdm16/code_segments.py
@@ -3,12 +3,12 @@ from typing import Optional
 
 import bitstruct
 
-from cocas.abstract_code_segments import CodeSegmentsInterface
-from cocas.ast_nodes import LabelNode, RegisterNode, RelocatableExpressionNode, TemplateFieldNode
-from cocas.code_block import Section
+from cocas.assembler.ast_nodes import LabelNode, RegisterNode, RelocatableExpressionNode, TemplateFieldNode
+from cocas.assembler.code_block import Section
 from cocas.error import CdmException, CdmExceptionTag
-from cocas.location import CodeLocation
 from cocas.object_module import ExternalEntry, ObjectSectionRecord
+from cocas.object_module.location import CodeLocation
+from cocas.targets import CodeSegmentsInterface
 
 
 def pack(fmt, *args):
