+++ cocas/targets/cdm8/target_instructions.py
@@ -4,10 +4,9 @@ from typing import Callable, get_args, get_origin
 
 from bitstruct import pack
 
-from cocas.abstract_code_segments import CodeSegmentsInterface
-from cocas.abstract_instructions import TargetInstructionsInterface
-from cocas.ast_nodes import InstructionNode, LabelNode, RegisterNode, RelocatableExpressionNode
+from cocas.assembler.ast_nodes import InstructionNode, LabelNode, RegisterNode, RelocatableExpressionNode
 from cocas.error import CdmException, CdmExceptionTag, CdmTempException
+from cocas.targets import CodeSegmentsInterface, TargetInstructionsInterface
 
 from .code_segments import CodeSegments
