+++ cocas/targets/cdm8e/target_instructions.py
@@ -2,10 +2,9 @@ from typing import Union, get_args, get_origin
 
 import bitstruct
 
-from cocas.abstract_code_segments import CodeSegmentsInterface
-from cocas.abstract_instructions import TargetInstructionsInterface
-from cocas.ast_nodes import InstructionNode, LabelNode, RegisterNode, RelocatableExpressionNode
+from cocas.assembler.ast_nodes import InstructionNode, LabelNode, RegisterNode, RelocatableExpressionNode
 from cocas.error import CdmException, CdmExceptionTag, CdmTempException
+from cocas.targets import CodeSegmentsInterface, TargetInstructionsInterface
 
 from .code_segments import CodeSegments
